pyproject.toml
[project]
name = "agent-chat-cli"
version = "0.2.15"
description = "Multi-Protocol (A2A/MCP) Agent Chat CLI Interface"
authors = [{name = "Sri Aradhyula", email = "sraradhy@cisco.com"}]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  "click>=8.2.1",
  "dotenv>=0.9.9",
  "httpx>=0.28.1",
  "httpx-sse>=0.4.0",
  "rich>=14.0.0",
  "agntcy-app-sdk>=0.1.0",
  "slim-bindings>=0.3.6",
  "a2a-sdk==0.2.16",
]

[project.scripts]
agent-chat-cli = "agent_chat_cli.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.13.3",
    "commitizen>=4.8.3",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_provider = "pep621"
version_files = [
    "agent_chat_cli/__init__.py:__version__"
]

